+++ .dialyzer_ignore.exs
@@ -1,4 +1,5 @@
 [
   {"test/support/test_scheduler.ex", :pattern_match, 0},
-  {"test/support/test_scheduler.ex", :pattern_match, 2}
+  {"test/support/test_scheduler.ex", :pattern_match, 2},
+  {"test/support/router.ex", :pattern_match, 16}
 ]